% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrBoost.R
\name{build_local_design}
\alias{build_local_design}
\title{Build Local Design Matrix}
\usage{
build_local_design(Xsub, local_model_type = "constant")
}
\arguments{
\item{Xsub}{numeric matrix (n_sub x p)}

\item{local_model_type}{in c("constant","linear","poly2")}
}
\value{
numeric matrix for local OLS fits
}
\description{
Creates a design matrix for local modeling
("constant", "linear", or "poly2").
}
\keyword{internal}
