% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_disjoint_rule_boost.R
\name{predict_disjoint_rule_boost}
\alias{predict_disjoint_rule_boost}
\title{Predict function for new Disjoint Rule Boosting objects}
\usage{
predict_disjoint_rule_boost(
  object,
  X_new,
  outcome_type = c("continuous", "binary"),
  logic = FALSE,
  type = c("link", "prob", "response")
)
}
\arguments{
\item{object}{the list returned by disjoint_rule_boost()}

\item{X_new}{matrix or data.frame of new feature values}

\item{outcome_type}{either "continuous" or "binary"
(we might store it in \code{object$outcome_type} if you like)}

\item{logic}{whether the rules are bounding-box or logic
(we could auto-detect if \code{rule_info$region_info$rule_string} is present)}

\item{type}{"link" (log-odds for binary, numeric for continuous),
"prob" => probability for binary,
"response" => 0/1 for binary.}
}
\value{
Numeric vector of predictions (length = nrow(X_new)).
}
\description{
Given the output of disjoint_rule_boost() (which returns
\code{init} + \code{rule_list}), we apply those rules to
new data X_new to compute predictions.
\itemize{
\item For outcome_type="continuous", the result is numeric predictions.
\item For outcome_type="binary", the result is log-odds by default;
use \code{type="prob"} to get probabilities, or
\code{type="response"} for 0/1 classification.
}
}
