% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrBoost.R
\name{predict_disjoint_rule_boost}
\alias{predict_disjoint_rule_boost}
\title{Predict function for Disjoint Rule Boosting objects
...}
\usage{
predict_disjoint_rule_boost(
  object,
  X_new,
  outcome_type = c("continuous", "binary"),
  type = c("link", "prob", "response")
)
}
\description{
Predict function for Disjoint Rule Boosting objects
...
}
